[Refactoring] Moved DeathLogic to own folder + renaming
