+++ Assets/Features/Player/Controller/CharacterInput/QueueInput.cs
@@ -1,0 +1,7 @@
+﻿namespace Features.Player.Controller.CharacterInput
+{
+    public class QueueInput
+    {
+        
+    }
+}
