+++ Assets/Features/Player/DeathLogic.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 37b842e26597cbb4e835a90eead38e73
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
